PEGTL Path Parser Extraction
- Separated out existing PEGTL path parser into its own set of header
  / source files in preparation for it to be used directly in
  a PEGTL based text file format parser
- Explicitly namespaced PEGTL types
